Add tests for some template functions.
